refactor(api): page the substrate reads the way the rest of the API pages
`common.proto` has carried `PageRequest`/`PageResponse` since before this branch,
and every other paged read takes them: ListCheckpoints, ListAgentInstances,
ListAgentInstanceShares, ListScheduledRuns, ListScheduledRunExecutions. The two
substrate list calls were the only ones spelling out `page_size`, `page_token` and
`next_page_token` themselves — down to re-declaring the `lte: 100` cap that
`PageRequest.limit` already carries, so the same rule was written in two places
and could disagree.

They now take `PageRequest page` and answer with `PageResponse page`, which also
puts the UI on the shape it already uses for conversations: `page: { limit,
pageToken }` out, `response.page.nextPageToken` back.

`SubstrateListInput.PageSize` becomes an `int` for the same reason — that is what
the agent-instance and checkpoint services take, with the int32 kept for ate-api's
own request where it belongs.

No behaviour changes: the cap, the default, and the refusal above the cap are
where they were, and `GetSubstrateSummary` keeps the counts, since `PageResponse`
carries a token and nothing else.

+++ go/core/internal/service/system/substrate.go
@@ -68,8 +68,9 @@ const maxATEPagesPerWalk = 10_000
 type SubstrateListInput struct {
 	// Empty means every namespace the controller observes.
 	Namespace string
-	// Zero means defaultSubstratePageSize.
-	PageSize int32
+	// Zero means defaultSubstratePageSize. An `int` as the other paged services take
+	// it; ate-api's own request is what wants the int32.
+	PageSize int
 	// Empty for the first page; otherwise the previous answer's NextPageToken.
 	PageToken string
 }
@@ -490,18 +491,18 @@ func allowedWorkerNamespace(namespace string, allowAll bool, allowed map[string]
 	return ok
 }
 
-func substratePageSize(requested int32) (int32, error) {
+func substratePageSize(requested int) (int32, error) {
 	switch {
 	case requested < 0:
 		return 0, serviceerrors.NewInvalidArgument(fmt.Sprintf("invalid page size %d: must not be negative", requested), nil)
 	case requested == 0:
 		return defaultSubstratePageSize, nil
-	case requested > maxSubstratePageSize:
+	case requested > int(maxSubstratePageSize):
 		return 0, serviceerrors.NewInvalidArgument(
 			fmt.Sprintf("invalid page size %d: the maximum is %d", requested, maxSubstratePageSize),
 			nil,
 		)
 	default:
-		return requested, nil
+		return int32(requested), nil
 	}
 }
